moving customEmojis interface to `definitions/IEmoji` file
