+++ app/lib/methods/getCustomEmojis.ts
@@ -1,12 +1,11 @@
 import orderBy from 'lodash/orderBy';
 import { sanitizedRaw } from '@nozbe/watermelondb/RawRecord';
 
-import { ICustomEmojis } from '../../reducers/customEmojis';
 import { store as reduxStore } from '../store/auxStore';
 import database from '../database';
 import log from '../../utils/log';
 import { setCustomEmojis as setCustomEmojisAction } from '../../actions/customEmojis';
-import { ICustomEmojiModel, TCustomEmojiModel } from '../../definitions';
+import { ICustomEmojiModel, TCustomEmojiModel, ICustomEmojis } from '../../definitions';
 import sdk from '../services/sdk';
 import { compareServerVersion } from './helpers/compareServerVersion';
